feat(pathfinder/sync/l2): download signature concurrently with the block
This change makes the signature download concurrent with the block download.
We start a task for the signature download right before even starting the
block download but only actually use the result once we have downloaded
the block.

There is a potential race condition here: the signature  download can
return with a "block not found" error even though the block download
succeeded. We can handle this case by retrying the the signature
download.